feat(sui-ssr): add option to only use new react context api
